@@ -5,7 +5,6 @@
 using Microsoft.Diagnostics.DebugServices;
 using Microsoft.Diagnostics.DebugServices.Implementation;
 using Microsoft.Diagnostics.ExtensionCommands;
-using Microsoft.Diagnostics.Repl;
 using Microsoft.Diagnostics.Runtime;
 using Microsoft.Diagnostics.Runtime.Utilities;
 using SOS.Hosting;
@@ -38,14 +37,13 @@ namespace SOS.Extensions
         internal DebuggerServices DebuggerServices { get; private set; }
 
         private readonly ServiceProvider _serviceProvider;
-        private readonly CommandProcessor _commandProcessor;
+        private readonly CommandService _commandService;
         private readonly SymbolService _symbolService;
         private readonly HostWrapper _hostWrapper;
         private ContextServiceFromDebuggerServices _contextService;
         private int _targetIdFactory;
         private ITarget _target;
         private TargetWrapper _targetWrapper;
-        private IMemoryService _memoryService;
 
         /// <summary>
         /// Enable the assembly resolver to get the right versions in the same directory as this assembly.
@@ -99,17 +97,16 @@ namespace SOS.Extensions
         {
             _serviceProvider = new ServiceProvider();
             _symbolService = new SymbolService(this);
-            _commandProcessor = new CommandProcessor(RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ? ">!ext" : null);
-            _commandProcessor.AddCommands(new Assembly[] { typeof(HostServices).Assembly });
-            _commandProcessor.AddCommands(new Assembly[] { typeof(ClrMDHelper).Assembly });
+            _commandService = new CommandService(RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ? ">!ext" : null);
+            _commandService.AddCommands(new Assembly[] { typeof(HostServices).Assembly });
+            _commandService.AddCommands(new Assembly[] { typeof(ClrMDHelper).Assembly });
 
             _serviceProvider.AddService<IHost>(this);
-            _serviceProvider.AddService<ICommandService>(_commandProcessor);
+            _serviceProvider.AddService<ICommandService>(_commandService);
             _serviceProvider.AddService<ISymbolService>(_symbolService);
 
             _hostWrapper = new HostWrapper(this, () => _targetWrapper);
-            _hostWrapper.AddServiceWrapper(IID_IHostServices, this);
-            _hostWrapper.AddServiceWrapper(SymbolServiceWrapper.IID_ISymbolService, () => new SymbolServiceWrapper(this, () => _memoryService));
+            _hostWrapper.ServiceWrapper.AddServiceWrapper(IID_IHostServices, this);
 
             VTableBuilder builder = AddInterface(IID_IHostServices, validate: false);
             builder.AddMethod(new GetHostDelegate(GetHost));
@@ -129,7 +126,7 @@ namespace SOS.Extensions
         protected override void Destroy()
         {
             Trace.TraceInformation("HostServices.Destroy");
-            _hostWrapper.RemoveServiceWrapper(IID_IHostServices);
+            _hostWrapper.ServiceWrapper.RemoveServiceWrapper(IID_IHostServices);
             _hostWrapper.Release();
         }
 
@@ -152,7 +149,6 @@ namespace SOS.Extensions
             if (target == _target)
             {
                 _target = null;
-                _memoryService = null;
                 if (_targetWrapper != null)
                 {
                     _targetWrapper.Release();
@@ -220,7 +216,7 @@ namespace SOS.Extensions
                 });
 
                 // Add each extension command to the native debugger
-                foreach ((string name, string help, IEnumerable<string> aliases) in _commandProcessor.Commands)
+                foreach ((string name, string help, IEnumerable<string> aliases) in _commandService.Commands)
                 {
                     hr = DebuggerServices.AddCommand(name, help, aliases);
                     if (hr != HResult.S_OK)
@@ -261,7 +257,7 @@ namespace SOS.Extensions
                 _target = new TargetFromDebuggerServices(DebuggerServices, this, _targetIdFactory++);
                 _contextService.SetCurrentTarget(_target);
                 _targetWrapper = new TargetWrapper(_contextService.Services);
-                _memoryService = _contextService.Services.GetService<IMemoryService>();
+                _targetWrapper.ServiceWrapper.AddServiceWrapper(SymbolServiceWrapper.IID_ISymbolService, () => new SymbolServiceWrapper(_symbolService, _target.Services.GetService<IMemoryService>()));
             }
             catch (Exception ex)
             {
@@ -325,7 +321,7 @@ namespace SOS.Extensions
             }
             try
             {
-                return _commandProcessor.Execute(commandLine, _contextService.Services);
+                return _commandService.Execute(commandLine, _contextService.Services);
             }
             catch (Exception ex)
             {
@@ -340,7 +336,7 @@ namespace SOS.Extensions
         {
             try
             {
-                if (!_commandProcessor.DisplayHelp(command, _contextService.Services))
+                if (!_commandService.DisplayHelp(command, _contextService.Services))
                 {
                     return HResult.E_INVALIDARG;
                 }
