@@ -5,8 +5,6 @@
 using Microsoft.Diagnostics.DebugServices;
 using Microsoft.Diagnostics.Runtime.Utilities;
 using System;
-using System.Collections.Generic;
-using System.Diagnostics;
 using System.Runtime.InteropServices;
 
 namespace SOS.Hosting
@@ -17,8 +15,8 @@ namespace SOS.Hosting
 
         private readonly IHost _host;
         private readonly Func<TargetWrapper> _getTarget;
-        private readonly Dictionary<Guid, Func<COMCallableIUnknown>> _factories = new Dictionary<Guid, Func<COMCallableIUnknown>>();
-        private readonly Dictionary<Guid, COMCallableIUnknown> _wrappers = new Dictionary<Guid, COMCallableIUnknown>();
+
+        public ServiceWrapper ServiceWrapper { get; } = new ServiceWrapper();
 
         public IntPtr IHost { get; }
 
@@ -29,73 +27,14 @@ namespace SOS.Hosting
 
             VTableBuilder builder = AddInterface(IID_IHost, validate: false);
             builder.AddMethod(new GetHostTypeDelegate(GetHostType));
-            builder.AddMethod(new GetServiceDelegate(GetService));
+            builder.AddMethod(new GetServiceDelegate(ServiceWrapper.GetService));
             builder.AddMethod(new GetCurrentTargetDelegate(GetCurrentTarget));
             IHost = builder.Complete();
 
             AddRef();
         }
 
-        protected override void Destroy()
-        {
-            Trace.TraceInformation("HostWrapper.Destroy");
-            foreach (var wrapper in _wrappers.Values)
-            {
-                wrapper.Release();
-            }
-            _wrappers.Clear();
-        }
-
-        /// <summary>
-        /// Add service instance factory
-        /// </summary>
-        /// <param name="serviceId">guid</param>
-        /// <param name="factory">factory delegate</param>
-        public void AddServiceWrapper(in Guid serviceId, Func<COMCallableIUnknown> factory)
-        {
-            _factories.Add(serviceId, factory);
-        }
-
-        /// <summary>
-        /// Add service instance
-        /// </summary>
-        /// <param name="serviceId">guid</param>
-        /// <param name="service">instance</param>
-        public void AddServiceWrapper(in Guid serviceId, COMCallableIUnknown service)
-        {
-            _wrappers.Add(serviceId, service);
-        }
-
-        /// <summary>
-        /// Remove the service instance
-        /// </summary>
-        /// <param name="serviceId">guid</param>
-        public void RemoveServiceWrapper(in Guid serviceId)
-        {
-            _factories.Remove(serviceId);
-            _wrappers.Remove(serviceId);
-        }
-
-        /// <summary>
-        /// Returns the wrapper instance for the guid
-        /// </summary>
-        /// <param name="serviceId">service guid</param>
-        /// <returns>instance or null</returns>
-        public COMCallableIUnknown GetServiceWrapper(in Guid serviceId)
-        {
-            if (!_wrappers.TryGetValue(serviceId, out COMCallableIUnknown service))
-            {
-                if (_factories.TryGetValue(serviceId, out Func<COMCallableIUnknown> factory))
-                {
-                    service = factory();
-                    if (service != null)
-                    {
-                        _wrappers.Add(serviceId, service);
-                    }
-                }
-            }
-            return service;
-        }
+        protected override void Destroy() => ServiceWrapper.Dispose();
 
         #region IHost
 
@@ -103,26 +42,6 @@ namespace SOS.Hosting
         /// Returns the host type
         /// </summary>
         private HostType GetHostType(IntPtr self) => _host.HostType;
-
-        /// <summary>
-        /// Returns the native service for the given interface id. There is 
-        /// only a limited set of services that can be queried through this
-        /// function. Adds a reference like QueryInterface.
-        /// </summary>
-        /// <param name="serviceId">guid of the service</param>
-        /// <param name="service">pointer to return service instance</param>
-        /// <returns>S_OK or E_NOINTERFACE</returns>
-        private HResult GetService(IntPtr self, in Guid guid, out IntPtr ptr)
-        {
-            ptr = IntPtr.Zero;
-
-            COMCallableIUnknown wrapper = GetServiceWrapper(guid);
-            if (wrapper == null) {
-                return HResult.E_NOINTERFACE;
-            }
-            wrapper.AddRef();
-            return COMHelper.QueryInterface(wrapper.IUnknownObject, guid, out ptr);
-        }
 
         /// <summary>
         /// Returns the current target wrapper or null
@@ -151,7 +70,7 @@ namespace SOS.Hosting
             [In] IntPtr self);
 
         [UnmanagedFunctionPointer(CallingConvention.Winapi)]
-        private delegate HResult GetServiceDelegate(
+        internal delegate HResult GetServiceDelegate(
             [In] IntPtr self,
             [In] in Guid guid,
             [Out] out IntPtr ptr);
