@@ -133,7 +133,11 @@ ISymbolService* Extensions::GetSymbolService()
 {
     if (m_pSymbolService == nullptr)
     {
-        GetHost()->GetService(__uuidof(ISymbolService), (void**)&m_pSymbolService);
+	    ITarget* target = GetTarget();
+        if (target != nullptr)
+        {
+            target->GetService(__uuidof(ISymbolService), (void**)&m_pSymbolService);
+        }
     }
     return m_pSymbolService;
 }
