@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Microsoft.Diagnostics.DebugServices;
+using Microsoft.Diagnostics.TestHelpers;
 
 namespace Microsoft.Diagnostics.DebugServices.UnitTests
 {
